suite/openstack: add --archive-upload-url
When uploading job results with --upload, it may be non-trivial to
figure out the public facing URL where the job will be visible. When
used via teuthology-openstack, it defaults to

    http://teuthology-logs.public.ceph.com

so that using --wait --upload --teardown displays the URLs of the
uploaded job results that are accessible even after the cluster is
gone.

Signed-off-by: Loic Dachary <loic@dachary.org>