openstack: image names are not unique
Image names are not unique, even within tenants, and if the user has
access to multiple tenants, you CAN get easily multiple images with the
same name.

In such circumstances, server creation fails as the name of the
specified image is not unique; pass the unique ID to the image instead.

If there are still multiple images with the same name, within a single
tenant, log a warning and pick the first one returned.

$ openstack image list -f json --long \
  --property name='teuthology-ubuntu-14.04' \
  --private
[{"ID": "c8b3...", "Name": "teuthology-ubuntu-14.04", "Owner": "186c..."},
 {"ID": "9656...", "Name": "teuthology-ubuntu-14.04", "Owner": "7aff..."}]

Signed-off-by: Robin H. Johnson <robin.johnson@dreamhost.com>